+++ src/configChangeListener.ts
@@ -1,7 +1,22 @@
 import * as vscode from "vscode";
-import { ConanAPI } from "./api/conan/conanAPI";
+import { ConanAPI, ConanExecutionMode } from "./api/conan/conanAPI";
 
 export function configChangeListener(event: vscode.ConfigurationChangeEvent, conanApi: ConanAPI) {
     // TODO: Check if only vsconan configuration is changed, otherwise do nothing. Removing overhead.
-    conanApi.setPythonInterpreter(vscode.workspace.getConfiguration("vsconan").get("general.pythonInterpreter")!);
+
+    let pythonInterpreter: string = vscode.workspace.getConfiguration("vsconan").get("general.pythonInterpreter")!;
+    let conanExecutable: string = vscode.workspace.getConfiguration("vsconan").get("general.conanExecutable")!;
+    let conanExecutionMode: string = vscode.workspace.getConfiguration("vsconan").get("general.conanExecutionMode")!;
+
+    conanApi.setPythonInterpreter(pythonInterpreter);
+    conanApi.setConanExecutable(conanExecutable);
+    if (conanExecutionMode === "pythonInterpreter") {
+        conanApi.switchExecutionMode(ConanExecutionMode.python);
+    }
+    else if (conanExecutionMode === "conanExecutable") {
+        conanApi.switchExecutionMode(ConanExecutionMode.conan);
+    }
+    else {
+        conanApi.switchExecutionMode(ConanExecutionMode.conan);
+    }
 }